ztp tools

+++ demos/ztp/tools/libvirt/delete-sno-ztp-vm.sh
@@ -5,6 +5,7 @@
 # Define variables
 VM_NAME="sno-ztp"
 NETWORK_NAME="sno-ztp"
+DISK_PATH="/var/lib/libvirt/images/sno-ztp.qcow2"
 
 check_command() {
     if [ $? -ne 0 ]; then
@@ -15,30 +16,45 @@ check_command() {
 
 # Step 1: Stop and undefine the VM
 echo "Stopping and undefining the VM: $VM_NAME"
-virsh dominfo $VM_NAME &> /dev/null
+sudo virsh dominfo $VM_NAME &> /dev/null
 if [ $? -eq 0 ]; then
     # VM exists, attempt to destroy and undefine it
-    virsh destroy $VM_NAME
-    check_command "virsh destroy"
+    if [ $(sudo virsh dominfo $VM_NAME | grep 'State:' | awk '{print $2}') == "running" ]; then
+        echo "VM $VM_NAME is running. Attempting to destroy it..."
+        sudo virsh destroy $VM_NAME
+        check_command "virsh destroy"
+    else
+        echo "VM $VM_NAME is not running"
+    fi
 
-    virsh undefine $VM_NAME
+    sudo virsh undefine $VM_NAME
     check_command "virsh undefine"
 else
     echo "VM $VM_NAME does not exist."
 fi
 
+if [ -f "$DISK_PATH" ]; then
+    echo "Disk file $DISK_PATH exists. Deleting it..."
+    sudo rm -f "$DISK_PATH"
+    check_command "rm -f $DISK_PATH"
+else
+    echo "Disk file $DISK_PATH does not exist."
+fi
+
 # Step 2: Destroy and undefine the network
 echo "Destroying and undefining the network: $NETWORK_NAME"
-virsh net-info $NETWORK_NAME &> /dev/null
+sudo virsh net-info $NETWORK_NAME &> /dev/null
 if [ $? -eq 0 ]; then
     # Network exists, attempt to destroy and undefine it
-    virsh net-destroy $NETWORK_NAME
+    sudo virsh net-destroy $NETWORK_NAME
     check_command "virsh net-destroy"
 
-    virsh net-undefine $NETWORK_NAME
+    sudo virsh net-undefine $NETWORK_NAME
     check_command "virsh net-undefine"
 else
     echo "Network $NETWORK_NAME does not exist."
 fi
+
+sudo nmcli con delete sno-ztp-br
 
 echo "Cleanup completed successfully."
